This adds support for another date format.
One that we found on another image. This is really a nit with exiftool.
It prints all dates in exif format. Meaning it makes it's own format
for things with time zones: yyyy:mm:dd hh:mm:ss+hh:mm

It also brings test coverage back up. I think the order of the output
may be non-deterministic. So we may need to do someting to make th
output consistent. We'll see.